Better onclick behavior
When the submit button is clicked, start with the current location based on location.protocol, location.host, and location.pathname, so that we lose any querystrings or other additional stuff on the current location, then add the url to be analyzed as the only querystring, and go to that new location. Also, better error handling in the XMLHttpRequest, resolves with the error so that it displays a little nicer on the page. Need to make this better still, so the user knows what action they should take next, and why they got that error.